requirements.txt
Django>=4.2.0
pyodbc>=4.0.39
celery>=5.3.0
redis>=4.5.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1